+++ core/src/commonMain/kotlin/dev/erikchristensen/islandtime/UtcOffset.kt
@@ -27,9 +27,9 @@ inline class UtcOffset internal constructor(
     ): T {
         val sign = if (totalSeconds.isNegative) -1 else 1
         val absTotalSeconds = totalSeconds.absoluteValue
-        val hours = (absTotalSeconds.value / SECONDS_PER_HOUR.toInt()).hours
-        val minutes = ((absTotalSeconds.value % SECONDS_PER_HOUR.toInt()) / SECONDS_PER_MINUTE.toInt()).minutes
-        val seconds = absTotalSeconds % SECONDS_PER_MINUTE.toInt()
+        val hours = (absTotalSeconds.value / SECONDS_PER_HOUR).hours
+        val minutes = ((absTotalSeconds.value % SECONDS_PER_HOUR) / SECONDS_PER_MINUTE).minutes
+        val seconds = absTotalSeconds % SECONDS_PER_MINUTE
 
         return action(sign, hours, minutes, seconds)
     }
@@ -40,9 +40,9 @@ inline class UtcOffset internal constructor(
     fun <T> toComponents(
         action: (hours: IntHours, minutes: IntMinutes, seconds: IntSeconds) -> T
     ): T {
-        val hours = (totalSeconds.value / SECONDS_PER_HOUR.toInt()).hours
-        val minutes = ((totalSeconds.value % SECONDS_PER_HOUR.toInt()) / SECONDS_PER_MINUTE.toInt()).minutes
-        val seconds = totalSeconds % SECONDS_PER_MINUTE.toInt()
+        val hours = (totalSeconds.value / SECONDS_PER_HOUR).hours
+        val minutes = ((totalSeconds.value % SECONDS_PER_HOUR) / SECONDS_PER_MINUTE).minutes
+        val seconds = totalSeconds % SECONDS_PER_MINUTE
 
         return action(hours, minutes, seconds)
     }
@@ -59,8 +59,8 @@ inline class UtcOffset internal constructor(
     }
 
     companion object {
-        val MAX_TOTAL_SECONDS = (18 * SECONDS_PER_HOUR.toInt()).seconds
-        val MIN_TOTAL_SECONDS = (-18 * SECONDS_PER_HOUR.toInt()).seconds
+        val MAX_TOTAL_SECONDS = (18 * SECONDS_PER_HOUR).seconds
+        val MIN_TOTAL_SECONDS = (-18 * SECONDS_PER_HOUR).seconds
 
         val MIN = UtcOffset(MAX_TOTAL_SECONDS)
         val MAX = UtcOffset(MIN_TOTAL_SECONDS)
